feat: bulk import spots (#35)

+++ src/modules/geo/GoActive.Modules.Geo.Application/Sketch/Create/CreateSketchCommand.cs
@@ -9,7 +9,7 @@ public sealed record CreateSketchCommand : ICommand<Result<SketchDto>>
     public required string Title { get; init; }
     public required IReadOnlyCollection<ActivityType> ActivityTypes { get; init; }
     public required SketchLocation Location { get; init; }
-    public double? Altitude { get; init; }
+    public float? Altitude { get; init; }
 
     public sealed record SketchLocation
     {
